\name{P}
 \docType{data}
 \alias{P}
 \title{Simulated null p-values from the uniform distribution.}
 \description{
 This data set is a simulated 200 x 100 matrix of null p-values where each of the 100 columns
 corresponds to a distinct study and each column contains 200 simulated p-values. 
 }
 \usage{P}
 \format{matrix}
 \keyword{datasets}
